Add uploadImage in Utils

+++ src/utils/uploadImage.ts
@@ -1,5 +1,5 @@
-import { v2 as cloudinary } from 'cloudinary';
-import fs from 'fs';
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME!,
@@ -12,17 +12,18 @@ export const uploadImageToCloudinary = async (localFilePath: string) => {
     if (!localFilePath) return null;
 
     const response = await cloudinary.uploader.upload(localFilePath, {
-      resource_type: 'auto',
+      resource_type: "auto",
     });
 
-    console.log('File uploaded to Cloudinary:', response.url);
+    console.log("File uploaded to Cloudinary:", response.secure_url);
 
-    fs.unlinkSync(localFilePath);
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
 
     return response;
   } catch (error) {
-    console.error('Cloudinary upload error:', error);
-      fs.unlinkSync(localFilePath);
+    console.error("Cloudinary upload error:", error);
 
     return null;
   }
@@ -31,10 +32,10 @@ export const uploadImageToCloudinary = async (localFilePath: string) => {
 export const deleteImageFromCloudinary = async (publicId: string) => {
   try {
     const result = await cloudinary.uploader.destroy(publicId);
-    console.log('Cloudinary image deleted:', publicId);
+    console.log("Cloudinary image deleted:", publicId);
     return result;
   } catch (error) {
-    console.error('Failed to delete Cloudinary image:', error);
+    console.error("Failed to delete Cloudinary image:", error);
     throw error;
   }
 };
